+++ src/test/resources/com/pestphp/pest/snapshotTesting/snapshotAssertionUseStatement.php
@@ -1,0 +1,3 @@
+<?php
+use function Spatie\Snapshots\assertMatchesJsonSnapshot;
+use function Spatie\Snapshots\assertMatchesSnapshot;